# frontend/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Trippy API",
    "version": "v1"
  },
  "paths": {
    "/api/chat": {
      "post": {
        "tags": [
          "chat"
        ],
        "operationId": "chat",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/actions/{actionId}": {
      "get": {
        "tags": [
          "chat"
        ],
        "operationId": "get_action",
        "parameters": [
          {
            "name": "actionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PendingActionRead"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingActionRead"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingActionRead"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/actions/{actionId}/approve": {
      "post": {
        "tags": [
          "chat"
        ],
        "operationId": "approve_action",
        "parameters": [
          {
            "name": "actionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PendingActionRead"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingActionRead"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingActionRead"
                }
              }
            }
          }
        }
      }
    },
    "/api/actions/{actionId}/deny": {
      "post": {
        "tags": [
          "chat"
        ],
        "operationId": "deny_action",
        "parameters": [
          {
            "name": "actionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PendingActionRead"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingActionRead"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingActionRead"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/items": {
      "post": {
        "tags": [
          "items"
        ],
        "operationId": "create_item",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemCreate"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemCreate"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ItemCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ItemRead"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemRead"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemRead"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/items/{itemId}": {
      "patch": {
        "tags": [
          "items"
        ],
        "operationId": "update_item",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemUpdate"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemUpdate"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ItemUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ItemRead"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemRead"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemRead"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "items"
        ],
        "operationId": "delete_item",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/items/reorder/{sectionId}": {
      "put": {
        "tags": [
          "items"
        ],
        "operationId": "reorder_items",
        "parameters": [
          {
            "name": "sectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemReorderRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemReorderRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ItemReorderRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ItemRead"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ItemRead"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ItemRead"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ChatRequest": {
        "required": [
          "message"
        ],
        "type": "object",
        "properties": {
          "message": {
            "minLength": 1,
            "type": "string"
          },
          "conversation_id": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ItemCreate": {
        "required": [
          "place_id",
          "section_id"
        ],
        "type": "object",
        "properties": {
          "section_id": {
            "type": "string",
            "format": "uuid"
          },
          "place_id": {
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "sequence": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "start_time": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ItemRead": {
        "required": [
          "description",
          "id",
          "place_id",
          "section_id",
          "sequence"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "section_id": {
            "type": "string",
            "format": "uuid"
          },
          "place_id": {
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "minLength": 1,
            "type": "string"
          },
          "sequence": {
            "type": "integer",
            "format": "int32"
          },
          "start_time": {
            "type": "string",
            "nullable": true
          },
          "place": {
            "$ref": "#/components/schemas/PlaceRead"
          }
        },
        "additionalProperties": false
      },
      "ItemReorderRequest": {
        "required": [
          "item_ids"
        ],
        "type": "object",
        "properties": {
          "item_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "additionalProperties": false
      },
      "ItemUpdate": {
        "type": "object",
        "properties": {
          "place_id": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "sequence": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "start_time": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PendingActionRead": {
        "required": [
          "conversation_id",
          "created_at",
          "id",
          "payload",
          "status",
          "summary",
          "tool_name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "conversation_id": {
            "minLength": 1,
            "type": "string"
          },
          "tool_name": {
            "minLength": 1,
            "type": "string"
          },
          "payload": { },
          "summary": {
            "minLength": 1,
            "type": "string"
          },
          "status": {
            "minLength": 1,
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "result": {
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PlaceRead": {
        "required": [
          "city",
          "description",
          "id",
          "latitude",
          "longitude",
          "name",
          "region",
          "type"
        ],
        "type": "object",
        "properties": {
          "id": {
            "minLength": 1,
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "type": {
            "minLength": 1,
            "type": "string"
          },
          "city": {
            "minLength": 1,
            "type": "string"
          },
          "region": {
            "minLength": 1,
            "type": "string"
          },
          "neighborhood": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "minLength": 1,
            "type": "string"
          },
          "latitude": {
            "type": "number",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "format": "double"
          },
          "hours": {
            "type": "string",
            "nullable": true
          },
          "duration_minutes": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "price_range": {
            "type": "string",
            "nullable": true
          },
          "rating": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "seasonal_notes": {
            "type": "string",
            "nullable": true
          },
          "booking_required": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  },
...
}